pyproject.toml
[project]
name = "nswf_detect_api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "James Westover", email = "james@westover.dev"},
]
dependencies = [
    "nudenet>=3.4.2",
    "fastapi[standard]>=0.116.1",
    "pytest>=8.4.1",
    "datasets>=4.0.0",
    "requests>=2.32.4",
    "pillow>=11.3.0",
    "fastapi-login>=1.10.3",
    "typer>=0.16.0",
    "python-dotenv>=1.1.1",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.42",
    "aiosqlite>=0.21.0",
]
requires-python = ">=3.12,<3.14"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.scripts]
dev = { env = { PYTHONPATH = "." }, cmd = "uvicorn app.main:app --reload --host 0.0.0.0 --port 6969" }
configure = "python scripts/configure.py"
serve = "uvicorn app.main:app --host 0.0.0.0 --port 6969"
test = { env = { PYTHONPATH = "." }, cmd = "pytest" }
fetch-data = "python tests/get_sample_data.py"
lock-matrix = { env = { PDM_IGNORE_ACTIVE_VENV = "1" }, cmd = "bash -lc 'pdm lock --python 3.13 --platform macos_arm64 && pdm lock --append --python 3.13 --platform macos_x86_64 && pdm lock --append --python 3.12 --platform manylinux_2_36_x86_64 && pdm lock --append --python 3.12 --platform manylinux_2_36_aarch64'" }
systemd-install = "python scripts/install_systemd.py"
# systemd helpers (Raspberry Pi / Linux)
service-status = "bash -lc 'systemctl status --no-pager nsfw_detect_api'"
service-logs = "bash -lc 'journalctl -u nsfw_detect_api -f -n 100'"
service-logs-all = "bash -lc 'journalctl -u nsfw_detect_api --no-pager -n 2000'"
service-restart = "bash -lc 'sudo systemctl restart nsfw_detect_api'"


[tool.pdm.build]
includes = ["app", "tests"]
excludes = ["tests/fixtures/*"]
